requirements.txt
dash==2.18.1
dash_ag_grid==31.2.0
langchain==0.3.4
langchain_core==0.3.13
langchain_groq==0.2.0
numpy==2.1.2
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.0.1
